requirements.txt
openai==1.6.1
python-dotenv=0.21.0
